requirements.txt
# Backend 의존성

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.1

# Vector DB
pinecone-client==3.0.0
langchain==0.1.10

# API Integration
requests==2.31.0
aiohttp==3.9.2

# Auth
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1

# Utils
python-dotenv==1.0.0
logging-json==0.6
async-timeout==4.0.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.25.2

# ML/AI Integration
google-generativeai==0.3.0
groq==0.4.2
anthropic==0.7.11
openai==1.6.1
ollama==0.1.0

# Image Processing
opencv-python==4.9.0.80
pillow==10.1.0
numpy==1.26.3
scikit-image==0.22.0

# Finance
yfinance==0.2.35
pandas==2.1.4
ta==0.10.2

# Telegram Bot
python-telegram-bot==20.5
telegram==0.0.1

# WebSocket
websockets==12.0

# Utilities
click==8.1.7
typer==0.9.0
rich==13.7.0
pyyaml==6.0.1
